+++ src/transform/__tests__/defer-test.ts
@@ -1,5 +1,9 @@
 import { assert, expect } from 'chai';
-import type { DocumentNode } from 'graphql';
+import type {
+  DocumentNode,
+  InitialIncrementalExecutionResult,
+  SubsequentIncrementalExecutionResult,
+} from 'graphql';
 import {
   GraphQLID,
   GraphQLList,
@@ -16,11 +20,7 @@ import { resolveOnNextTick } from '../../__testUtils__/resolveOnNextTick.js';
 
 import { promiseWithResolvers } from '../../jsutils/promiseWithResolvers.js';
 
-import type {
-  LegacyInitialIncrementalExecutionResult,
-  LegacySubsequentIncrementalExecutionResult,
-} from '../legacyExecuteIncrementally.js';
-import { legacyExecuteIncrementally as execute } from '../legacyExecuteIncrementally.js';
+import { execute } from './execute.js';
 
 const friendType = new GraphQLObjectType({
   fields: {
@@ -149,8 +149,7 @@ async function complete(
 
   if ('initialResult' in result) {
     const results: Array<
-      | LegacyInitialIncrementalExecutionResult
-      | LegacySubsequentIncrementalExecutionResult
+      InitialIncrementalExecutionResult | SubsequentIncrementalExecutionResult
     > = [result.initialResult];
     for await (const patch of result.subsequentResults) {
       results.push(patch);
@@ -160,7 +159,7 @@ async function complete(
   return result;
 }
 
-describe('Execute: legacy defer directive format', () => {
+describe('Execute: defer directive', () => {
   it('Can defer fragments containing scalar types', async () => {
     const document = parse(`
       query HeroNameQuery {
@@ -182,6 +181,7 @@ describe('Execute: legacy defer directive format', () => {
             id: '1',
           },
         },
+        pending: [{ id: '0', path: ['hero'] }],
         hasNext: true,
       },
       {
@@ -190,9 +190,10 @@ describe('Execute: legacy defer directive format', () => {
             data: {
               name: 'Luke',
             },
-            path: ['hero'],
+            id: '0',
           },
         ],
+        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -236,15 +237,17 @@ describe('Execute: legacy defer directive format', () => {
     expectJSON(result).toDeepEqual([
       {
         data: { hero: { id: '1' } },
+        pending: [{ id: '0', path: ['hero'] }],
         hasNext: true,
       },
       {
         incremental: [
           {
             data: { name: 'Luke' },
-            path: ['hero'],
+            id: '0',
           },
         ],
+        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -285,6 +288,7 @@ describe('Execute: legacy defer directive format', () => {
             id: '1',
           },
         },
+        pending: [{ id: '0', path: ['hero'] }],
         hasNext: true,
       },
       {
@@ -293,9 +297,10 @@ describe('Execute: legacy defer directive format', () => {
             data: {
               name: 'Luke',
             },
-            path: ['hero'],
+            id: '0',
           },
         ],
+        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -341,6 +346,7 @@ describe('Execute: legacy defer directive format', () => {
             id: '1',
           },
         },
+        pending: [{ id: '0', path: ['hero'] }],
         hasNext: true,
       },
       {
@@ -349,9 +355,10 @@ describe('Execute: legacy defer directive format', () => {
             data: {
               name: 'Luke',
             },
-            path: ['hero'],
+            id: '0',
           },
         ],
+        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -373,6 +380,7 @@ describe('Execute: legacy defer directive format', () => {
     expectJSON(result).toDeepEqual([
       {
         data: {},
+        pending: [{ id: '0', path: [], label: 'DeferQuery' }],
         hasNext: true,
       },
       {
@@ -383,10 +391,10 @@ describe('Execute: legacy defer directive format', () => {
                 id: '1',
               },
             },
-            path: [],
-            label: 'DeferQuery',
+            id: '0',
           },
         ],
+        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -414,6 +422,7 @@ describe('Execute: legacy defer directive format', () => {
     expectJSON(result).toDeepEqual([
       {
         data: {},
+        pending: [{ id: '0', path: [], label: 'DeferQuery' }],
         hasNext: true,
       },
       {
@@ -431,10 +440,10 @@ describe('Execute: legacy defer directive format', () => {
                 path: ['hero', 'name'],
               },
             ],
-            path: [],
-            label: 'DeferQuery',
+            id: '0',
           },
         ],
+        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -463,25 +472,26 @@ describe('Execute: legacy defer directive format', () => {
         data: {
           hero: {},
         },
+        pending: [{ id: '0', path: ['hero'], label: 'DeferTop' }],
         hasNext: true,
       },
       {
+        pending: [{ id: '1', path: ['hero'], label: 'DeferNested' }],
         incremental: [
           {
             data: {
               id: '1',
             },
-            path: ['hero'],
-            label: 'DeferTop',
+            id: '0',
           },
           {
             data: {
               friends: [{ name: 'Han' }, { name: 'Leia' }, { name: 'C-3PO' }],
             },
-            path: ['hero'],
-            label: 'DeferNested',
+            id: '1',
           },
         ],
+        completed: [{ id: '0' }, { id: '1' }],
         hasNext: false,
       },
     ]);
@@ -499,28 +509,13 @@ describe('Execute: legacy defer directive format', () => {
       }
     `);
     const result = await complete(document);
-    expectJSON(result).toDeepEqual([
-      {
-        data: {
-          hero: {
-            name: 'Luke',
-          },
+    expectJSON(result).toDeepEqual({
+      data: {
+        hero: {
+          name: 'Luke',
         },
-        hasNext: true,
       },
-      {
-        incremental: [
-          {
-            data: {
-              name: 'Luke',
-            },
-            path: ['hero'],
-            label: 'DeferTop',
-          },
-        ],
-        hasNext: false,
-      },
-    ]);
+    });
   });
   it('Can defer a fragment that is also not deferred, non-deferred fragment is first', async () => {
     const document = parse(`
@@ -560,12 +555,12 @@ describe('Execute: legacy defer directive format', () => {
     expectJSON(result).toDeepEqual([
       {
         data: { hero: { id: '1' } },
+        pending: [{ id: '0', path: ['hero'], label: 'InlineDeferred' }],
         hasNext: true,
       },
       {
-        incremental: [
-          { data: { name: 'Luke' }, path: ['hero'], label: 'InlineDeferred' },
-        ],
+        incremental: [{ data: { name: 'Luke' }, id: '0' }],
+        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -610,10 +605,12 @@ describe('Execute: legacy defer directive format', () => {
         data: {
           hero: {},
         },
+        pending: [{ id: '0', path: ['hero'] }],
         hasNext: true,
       },
       {
-        incremental: [{ data: { name: 'Luke' }, path: ['hero'] }],
+        incremental: [{ data: { name: 'Luke' }, id: '0' }],
+        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -638,6 +635,10 @@ describe('Execute: legacy defer directive format', () => {
         data: {
           hero: {},
         },
+        pending: [
+          { id: '0', path: ['hero'], label: 'DeferID' },
+          { id: '1', path: ['hero'], label: 'DeferName' },
+        ],
         hasNext: true,
       },
       {
@@ -646,17 +647,16 @@ describe('Execute: legacy defer directive format', () => {
             data: {
               id: '1',
             },
-            path: ['hero'],
-            label: 'DeferID',
+            id: '0',
           },
           {
             data: {
               name: 'Luke',
             },
-            path: ['hero'],
-            label: 'DeferName',
+            id: '1',
           },
         ],
+        completed: [{ id: '0' }, { id: '1' }],
         hasNext: false,
       },
     ]);
@@ -681,21 +681,30 @@ describe('Execute: legacy defer directive format', () => {
     expectJSON(result).toDeepEqual([
       {
         data: {},
+        pending: [
+          { id: '0', path: [], label: 'DeferID' },
+          { id: '1', path: [], label: 'DeferName' },
+        ],
         hasNext: true,
       },
       {
         incremental: [
           {
-            data: { hero: { id: '1' } },
-            path: [],
-            label: 'DeferID',
+            data: { hero: {} },
+            id: '0',
           },
           {
-            data: { hero: { name: 'Luke' } },
-            path: [],
-            label: 'DeferName',
+            data: { id: '1' },
+            id: '0',
+            subPath: ['hero'],
+          },
+          {
+            data: { name: 'Luke' },
+            id: '1',
+            subPath: ['hero'],
           },
         ],
+        completed: [{ id: '0' }, { id: '1' }],
         hasNext: false,
       },
     ]);
@@ -725,21 +734,30 @@ describe('Execute: legacy defer directive format', () => {
     expectJSON(result).toDeepEqual([
       {
         data: {},
+        pending: [
+          { id: '0', path: [], label: 'DeferID' },
+          { id: '1', path: [], label: 'DeferName' },
+        ],
         hasNext: true,
       },
       {
         incremental: [
           {
-            data: { hero: { id: '1' } },
-            path: [],
-            label: 'DeferID',
+            data: { hero: {} },
+            id: '0',
           },
           {
-            data: { hero: { name: 'Luke' } },
-            path: [],
-            label: 'DeferName',
+            data: { id: '1' },
+            id: '0',
+            subPath: ['hero'],
+          },
+          {
+            data: { name: 'Luke' },
+            id: '1',
+            subPath: ['hero'],
           },
         ],
+        completed: [{ id: '0' }, { id: '1' }],
         hasNext: false,
       },
     ]);
@@ -766,6 +784,10 @@ describe('Execute: legacy defer directive format', () => {
         data: {
           hero: {},
         },
+        pending: [
+          { id: '0', path: ['hero'], label: 'DeferID' },
+          { id: '1', path: [], label: 'DeferName' },
+        ],
         hasNext: true,
       },
       {
@@ -774,17 +796,17 @@ describe('Execute: legacy defer directive format', () => {
             data: {
               id: '1',
             },
-            path: ['hero'],
-            label: 'DeferID',
+            id: '0',
           },
           {
             data: {
-              hero: { name: 'Luke' },
+              name: 'Luke',
             },
-            path: [],
-            label: 'DeferName',
+            id: '1',
+            subPath: ['hero'],
           },
         ],
+        completed: [{ id: '0' }, { id: '1' }],
         hasNext: false,
       },
     ]);
@@ -807,9 +829,11 @@ describe('Execute: legacy defer directive format', () => {
     expectJSON(result).toDeepEqual([
       {
         data: {},
+        pending: [{ id: '0', path: [], label: 'DeferName' }],
         hasNext: true,
       },
       {
+        pending: [{ id: '1', path: ['hero'], label: 'DeferID' }],
         incremental: [
           {
             data: {
@@ -817,17 +841,16 @@ describe('Execute: legacy defer directive format', () => {
                 name: 'Luke',
               },
             },
-            path: [],
-            label: 'DeferName',
+            id: '0',
           },
           {
             data: {
               id: '1',
             },
-            path: ['hero'],
-            label: 'DeferID',
+            id: '1',
           },
         ],
+        completed: [{ id: '0' }, { id: '1' }],
         hasNext: false,
       },
     ]);
@@ -888,6 +911,10 @@ describe('Execute: legacy defer directive format', () => {
     const result1 = executeResult.initialResult;
     expectJSON(result1).toDeepEqual({
       data: {},
+      pending: [
+        { id: '0', path: [] },
+        { id: '1', path: [] },
+      ],
       hasNext: true,
     });
 
@@ -899,16 +926,23 @@ describe('Execute: legacy defer directive format', () => {
     const result2 = await iterator.next();
     expectJSON(result2).toDeepEqual({
       value: {
+        pending: [{ id: '2', path: ['a'] }],
         incremental: [
           {
             data: { a: {} },
-            path: [],
+            id: '0',
           },
           {
-            data: { b: { c: { d: 'd' } } },
-            path: ['a'],
+            data: { b: {} },
+            id: '2',
+          },
+          {
+            data: { c: { d: 'd' } },
+            id: '2',
+            subPath: ['b'],
           },
         ],
+        completed: [{ id: '0' }, { id: '2' }],
         hasNext: true,
       },
       done: false,
@@ -922,18 +956,20 @@ describe('Execute: legacy defer directive format', () => {
     const result3 = await iterator.next();
     expectJSON(result3).toDeepEqual({
       value: {
+        pending: [{ id: '3', path: ['a'] }],
         incremental: [
           {
-            data: { a: { someField: 'someField' } },
-            path: [],
+            data: { someField: 'someField' },
+            id: '1',
+            subPath: ['a'],
           },
           {
-            data: {
-              b: { e: { f: 'f' } },
-            },
-            path: ['a'],
+            data: { e: { f: 'f' } },
+            id: '3',
+            subPath: ['b'],
           },
         ],
+        completed: [{ id: '1' }, { id: '3' }],
         hasNext: false,
       },
       done: false,
@@ -1004,6 +1040,10 @@ describe('Execute: legacy defer directive format', () => {
     const result1 = executeResult.initialResult;
     expectJSON(result1).toDeepEqual({
       data: {},
+      pending: [
+        { id: '0', path: [] },
+        { id: '1', path: [] },
+      ],
       hasNext: true,
     });
 
@@ -1015,16 +1055,17 @@ describe('Execute: legacy defer directive format', () => {
     const result2 = await iterator.next();
     expectJSON(result2).toDeepEqual({
       value: {
+        pending: [
+          { id: '2', path: ['a'] },
+          { id: '3', path: ['a'] },
+        ],
         incremental: [
           {
             data: { a: {} },
-            path: [],
-          },
-          {
-            data: { a: {} },
-            path: [],
+            id: '0',
           },
         ],
+        completed: [{ id: '0' }, { id: '1' }],
         hasNext: true,
       },
       done: false,
@@ -1041,13 +1082,15 @@ describe('Execute: legacy defer directive format', () => {
         incremental: [
           {
             data: { b: { c: { d: 'd' } } },
-            path: ['a'],
+            id: '2',
           },
           {
-            data: { b: { c: { d: 'd' }, e: { f: 'f' } } },
-            path: ['a'],
+            data: { e: { f: 'f' } },
+            id: '3',
+            subPath: ['b'],
           },
         ],
+        completed: [{ id: '2' }, { id: '3' }],
         hasNext: false,
       },
       done: false,
@@ -1060,7 +1103,7 @@ describe('Execute: legacy defer directive format', () => {
     });
   });
 
-  it('Handles multiple defers on the same object', async () => {
+  it('Can deduplicate multiple defers on the same object', async () => {
     const document = parse(`
       query {
         hero {
@@ -1091,20 +1134,26 @@ describe('Execute: legacy defer directive format', () => {
     expectJSON(result).toDeepEqual([
       {
         data: { hero: { friends: [{}, {}, {}] } },
+        pending: [
+          { id: '0', path: ['hero', 'friends', 0] },
+          { id: '1', path: ['hero', 'friends', 1] },
+          { id: '2', path: ['hero', 'friends', 2] },
+        ],
         hasNext: true,
       },
       {
         incremental: [
-          { data: { id: '2', name: 'Han' }, path: ['hero', 'friends', 0] },
-          { data: { id: '3', name: 'Leia' }, path: ['hero', 'friends', 1] },
-          { data: { id: '4', name: 'C-3PO' }, path: ['hero', 'friends', 2] },
+          { data: { id: '2', name: 'Han' }, id: '0' },
+          { data: { id: '3', name: 'Leia' }, id: '1' },
+          { data: { id: '4', name: 'C-3PO' }, id: '2' },
         ],
+        completed: [{ id: '0' }, { id: '1' }, { id: '2' }],
         hasNext: false,
       },
     ]);
   });
 
-  it('Handles overlapping fields present in the initial payload', async () => {
+  it('Deduplicates fields present in the initial payload', async () => {
     const document = parse(`
       query {
         hero {
@@ -1155,32 +1204,24 @@ describe('Execute: legacy defer directive format', () => {
             },
           },
         },
+        pending: [{ id: '0', path: ['hero'] }],
         hasNext: true,
       },
       {
         incremental: [
           {
-            data: {
-              nestedObject: {
-                deeperObject: {
-                  bar: 'bar',
-                },
-              },
-              anotherNestedObject: {
-                deeperObject: {
-                  foo: 'foo',
-                },
-              },
-            },
-            path: ['hero'],
+            data: { bar: 'bar' },
+            id: '0',
+            subPath: ['nestedObject', 'deeperObject'],
           },
         ],
+        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
   });
 
-  it('Handles overlapping fields present in a parent defer payload', async () => {
+  it('Deduplicates fields present in a parent defer payload', async () => {
     const document = parse(`
       query {
         hero {
@@ -1206,9 +1247,11 @@ describe('Execute: legacy defer directive format', () => {
         data: {
           hero: {},
         },
+        pending: [{ id: '0', path: ['hero'] }],
         hasNext: true,
       },
       {
+        pending: [{ id: '1', path: ['hero', 'nestedObject', 'deeperObject'] }],
         incremental: [
           {
             data: {
@@ -1216,19 +1259,22 @@ describe('Execute: legacy defer directive format', () => {
                 deeperObject: { foo: 'foo' },
               },
             },
-            path: ['hero'],
+            id: '0',
           },
           {
-            data: { foo: 'foo', bar: 'bar' },
-            path: ['hero', 'nestedObject', 'deeperObject'],
+            data: {
+              bar: 'bar',
+            },
+            id: '1',
           },
         ],
+        completed: [{ id: '0' }, { id: '1' }],
         hasNext: false,
       },
     ]);
   });
 
-  it('Handles overlapping fields with deferred fragments at multiple levels', async () => {
+  it('Deduplicates fields with deferred fragments at multiple levels', async () => {
     const document = parse(`
       query {
         hero {
@@ -1279,31 +1325,37 @@ describe('Execute: legacy defer directive format', () => {
             },
           },
         },
+        pending: [{ id: '0', path: ['hero'] }],
         hasNext: true,
       },
       {
+        pending: [
+          { id: '1', path: ['hero', 'nestedObject'] },
+          { id: '2', path: ['hero', 'nestedObject', 'deeperObject'] },
+        ],
         incremental: [
           {
-            data: {
-              nestedObject: { deeperObject: { foo: 'foo', bar: 'bar' } },
-            },
-            path: ['hero'],
+            data: { bar: 'bar' },
+            id: '0',
+            subPath: ['nestedObject', 'deeperObject'],
           },
           {
-            data: { deeperObject: { foo: 'foo', bar: 'bar', baz: 'baz' } },
-            path: ['hero', 'nestedObject'],
+            data: { baz: 'baz' },
+            id: '1',
+            subPath: ['deeperObject'],
           },
           {
-            data: { foo: 'foo', bar: 'bar', baz: 'baz', bak: 'bak' },
-            path: ['hero', 'nestedObject', 'deeperObject'],
+            data: { bak: 'bak' },
+            id: '2',
           },
         ],
+        completed: [{ id: '0' }, { id: '1' }, { id: '2' }],
         hasNext: false,
       },
     ]);
   });
 
-  it('Handles overlapping fields from deferred fragments from different branches occurring at the same level', async () => {
+  it('Deduplicates multiple fields from deferred fragments from different branches occurring at the same level', async () => {
     const document = parse(`
       query {
         hero {
@@ -1339,16 +1391,9 @@ describe('Execute: legacy defer directive format', () => {
             },
           },
         },
-        hasNext: true,
-      },
-      {
-        incremental: [
-          {
-            data: {
-              nestedObject: { deeperObject: {} },
-            },
-            path: ['hero'],
-          },
+        pending: [
+          { id: '0', path: ['hero', 'nestedObject', 'deeperObject'] },
+          { id: '1', path: ['hero', 'nestedObject', 'deeperObject'] },
         ],
         hasNext: true,
       },
@@ -1358,22 +1403,22 @@ describe('Execute: legacy defer directive format', () => {
             data: {
               foo: 'foo',
             },
-            path: ['hero', 'nestedObject', 'deeperObject'],
+            id: '0',
           },
           {
             data: {
-              foo: 'foo',
               bar: 'bar',
             },
-            path: ['hero', 'nestedObject', 'deeperObject'],
+            id: '1',
           },
         ],
+        completed: [{ id: '0' }, { id: '1' }],
         hasNext: false,
       },
     ]);
   });
 
-  it('Handles deferred fragments in different branches at multiple non-overlapping levels', async () => {
+  it('Deduplicate fields with deferred fragments in different branches at multiple non-overlapping levels', async () => {
     const document = parse(`
       query {
         a {
@@ -1422,25 +1467,30 @@ describe('Execute: legacy defer directive format', () => {
             },
           },
         },
+        pending: [
+          { id: '0', path: ['a', 'b'] },
+          { id: '1', path: [] },
+        ],
         hasNext: true,
       },
       {
         incremental: [
           {
             data: { e: { f: 'f' } },
-            path: ['a', 'b'],
+            id: '0',
           },
           {
-            data: { a: { b: { e: { f: 'f' } } }, g: { h: 'h' } },
-            path: [],
+            data: { g: { h: 'h' } },
+            id: '1',
           },
         ],
+        completed: [{ id: '0' }, { id: '1' }],
         hasNext: false,
       },
     ]);
   });
 
-  it('Handles varying subfields with overlapping fields', async () => {
+  it('Correctly bundles varying subfields into incremental data records unique by defer combination, ignoring fields in a fragment masked by a parent defer', async () => {
     const document = parse(`
       query HeroNameQuery {
         ... @defer {
@@ -1463,30 +1513,33 @@ describe('Execute: legacy defer directive format', () => {
     expectJSON(result).toDeepEqual([
       {
         data: {},
+        pending: [
+          { id: '0', path: [] },
+          { id: '1', path: [] },
+        ],
         hasNext: true,
       },
       {
         incremental: [
           {
-            data: { hero: { id: '1' } },
-            path: [],
+            data: { hero: {} },
+            id: '0',
+          },
+          {
+            data: { id: '1' },
+            id: '0',
+            subPath: ['hero'],
           },
           {
             data: {
-              hero: {
-                name: 'Luke',
-                shouldBeWithNameDespiteAdditionalDefer: 'Luke',
-              },
-            },
-            path: [],
-          },
-          {
-            data: {
+              name: 'Luke',
               shouldBeWithNameDespiteAdditionalDefer: 'Luke',
             },
-            path: ['hero'],
+            id: '1',
+            subPath: ['hero'],
           },
         ],
+        completed: [{ id: '0' }, { id: '1' }],
         hasNext: false,
       },
     ]);
@@ -1524,12 +1577,27 @@ describe('Execute: legacy defer directive format', () => {
         data: {
           a: {},
         },
+        pending: [
+          { id: '0', path: [] },
+          { id: '1', path: ['a'] },
+        ],
         hasNext: true,
       },
       {
         incremental: [
           {
-            data: null,
+            data: { b: { c: {} } },
+            id: '1',
+          },
+          {
+            data: { d: 'd' },
+            id: '1',
+            subPath: ['b', 'c'],
+          },
+        ],
+        completed: [
+          {
+            id: '0',
             errors: [
               {
                 message:
@@ -1538,12 +1606,8 @@ describe('Execute: legacy defer directive format', () => {
                 path: ['a', 'b', 'c', 'nonNullErrorField'],
               },
             ],
-            path: [],
           },
-          {
-            data: { b: { c: { d: 'd' } } },
-            path: ['a'],
-          },
+          { id: '1' },
         ],
         hasNext: false,
       },
@@ -1585,16 +1649,28 @@ describe('Execute: legacy defer directive format', () => {
         data: {
           a: {},
         },
+        pending: [
+          { id: '0', path: [] },
+          { id: '1', path: ['a'] },
+        ],
         hasNext: true,
       },
       {
         incremental: [
           {
-            data: { a: { b: { c: { d: 'd' } } } },
-            path: [],
+            data: { b: { c: {} } },
+            id: '0',
+            subPath: ['a'],
           },
           {
-            data: null,
+            data: { d: 'd' },
+            id: '0',
+            subPath: ['a', 'b', 'c'],
+          },
+        ],
+        completed: [
+          {
+            id: '1',
             errors: [
               {
                 message:
@@ -1603,8 +1679,8 @@ describe('Execute: legacy defer directive format', () => {
                 path: ['a', 'b', 'c', 'nonNullErrorField'],
               },
             ],
-            path: ['a'],
           },
+          { id: '0' },
         ],
         hasNext: false,
       },
@@ -1642,12 +1718,16 @@ describe('Execute: legacy defer directive format', () => {
     expectJSON(result).toDeepEqual([
       {
         data: {},
+        pending: [
+          { id: '0', path: [] },
+          { id: '1', path: [] },
+        ],
         hasNext: true,
       },
       {
-        incremental: [
+        completed: [
           {
-            data: null,
+            id: '0',
             errors: [
               {
                 message:
@@ -1656,10 +1736,9 @@ describe('Execute: legacy defer directive format', () => {
                 path: ['a', 'b', 'c', 'someError'],
               },
             ],
-            path: [],
           },
           {
-            data: null,
+            id: '1',
             errors: [
               {
                 message:
@@ -1668,7 +1747,6 @@ describe('Execute: legacy defer directive format', () => {
                 path: ['a', 'b', 'c', 'anotherError'],
               },
             ],
-            path: [],
           },
         ],
         hasNext: false,
@@ -1713,12 +1791,32 @@ describe('Execute: legacy defer directive format', () => {
     expectJSON(result).toDeepEqual([
       {
         data: {},
+        pending: [
+          { id: '0', path: [] },
+          { id: '1', path: [] },
+        ],
         hasNext: true,
       },
       {
         incremental: [
           {
-            data: null,
+            data: { a: { b: { someC: {}, anotherC: {} } } },
+            id: '0',
+          },
+          {
+            data: { d: 'd' },
+            id: '0',
+            subPath: ['a', 'b', 'someC'],
+          },
+          {
+            data: { d: 'd' },
+            id: '0',
+            subPath: ['a', 'b', 'anotherC'],
+          },
+        ],
+        completed: [
+          {
+            id: '1',
             errors: [
               {
                 message:
@@ -1727,19 +1825,15 @@ describe('Execute: legacy defer directive format', () => {
                 path: ['a', 'b', 'someC', 'someError'],
               },
             ],
-            path: [],
           },
-          {
-            data: { a: { b: { someC: { d: 'd' }, anotherC: { d: 'd' } } } },
-            path: [],
-          },
+          { id: '0' },
         ],
         hasNext: false,
       },
     ]);
   });
 
-  it('handles a payload with a null that cannot be merged', async () => {
+  it('filters a payload with a null that cannot be merged', async () => {
     const document = parse(`
       query {
         ... @defer {
@@ -1786,21 +1880,31 @@ describe('Execute: legacy defer directive format', () => {
         data: {
           a: {},
         },
-        hasNext: true,
-      },
-      {
-        incremental: [
-          {
-            data: { b: { c: { d: 'd' } } },
-            path: ['a'],
-          },
+        pending: [
+          { id: '0', path: [] },
+          { id: '1', path: ['a'] },
         ],
         hasNext: true,
       },
       {
         incremental: [
           {
-            data: null,
+            data: { b: { c: {} } },
+            id: '1',
+          },
+          {
+            data: { d: 'd' },
+            id: '1',
+            subPath: ['b', 'c'],
+          },
+        ],
+        completed: [{ id: '1' }],
+        hasNext: true,
+      },
+      {
+        completed: [
+          {
+            id: '0',
             errors: [
               {
                 message:
@@ -1809,7 +1913,6 @@ describe('Execute: legacy defer directive format', () => {
                 path: ['a', 'b', 'c', 'nonNullErrorField'],
               },
             ],
-            path: [],
           },
         ],
         hasNext: false,
@@ -1840,41 +1943,19 @@ describe('Execute: legacy defer directive format', () => {
       },
       true,
     );
-    expectJSON(result).toDeepEqual([
-      {
-        data: {
-          hero: null,
+    expectJSON(result).toDeepEqual({
+      data: {
+        hero: null,
+      },
+      errors: [
+        {
+          message:
+            'Cannot return null for non-nullable field Hero.nonNullName.',
+          locations: [{ line: 4, column: 11 }],
+          path: ['hero', 'nonNullName'],
         },
-        errors: [
-          {
-            message:
-              'Cannot return null for non-nullable field Hero.nonNullName.',
-            locations: [{ line: 4, column: 11 }],
-            path: ['hero', 'nonNullName'],
-          },
-        ],
-        hasNext: true,
-      },
-      {
-        incremental: [
-          {
-            data: {
-              hero: null,
-            },
-            path: [],
-            errors: [
-              {
-                message:
-                  'Cannot return null for non-nullable field Hero.nonNullName.',
-                locations: [{ line: 4, column: 11 }],
-                path: ['hero', 'nonNullName'],
-              },
-            ],
-          },
-        ],
-        hasNext: false,
-      },
-    ]);
+      ],
+    });
   });
 
   it('Cancels deferred fields when deferred result exhibits null bubbling', async () => {
@@ -1901,6 +1982,7 @@ describe('Execute: legacy defer directive format', () => {
     expectJSON(result).toDeepEqual([
       {
         data: {},
+        pending: [{ id: '0', path: [] }],
         hasNext: true,
       },
       {
@@ -1917,15 +1999,16 @@ describe('Execute: legacy defer directive format', () => {
                 path: ['hero', 'nonNullName'],
               },
             ],
-            path: [],
+            id: '0',
           },
         ],
+        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
   });
 
-  it('Handles duplicated list fields', async () => {
+  it('Deduplicates list fields', async () => {
     const document = parse(`
       query {
         hero {
@@ -1941,30 +2024,16 @@ describe('Execute: legacy defer directive format', () => {
       }
     `);
     const result = await complete(document);
-    expectJSON(result).toDeepEqual([
-      {
-        data: {
-          hero: {
-            friends: [{ name: 'Han' }, { name: 'Leia' }, { name: 'C-3PO' }],
-          },
+    expectJSON(result).toDeepEqual({
+      data: {
+        hero: {
+          friends: [{ name: 'Han' }, { name: 'Leia' }, { name: 'C-3PO' }],
         },
-        hasNext: true,
       },
-      {
-        incremental: [
-          {
-            data: {
-              friends: [{ name: 'Han' }, { name: 'Leia' }, { name: 'C-3PO' }],
-            },
-            path: ['hero'],
-          },
-        ],
-        hasNext: false,
-      },
-    ]);
+    });
   });
 
-  it('Handles duplicated async iterable list fields', async () => {
+  it('Deduplicates async iterable list fields', async () => {
     const document = parse(`
       query {
         hero {
@@ -1987,24 +2056,12 @@ describe('Execute: legacy defer directive format', () => {
         },
       },
     });
-    expectJSON(result).toDeepEqual([
-      {
-        data: { hero: { friends: [{ name: 'Han' }] } },
-        hasNext: true,
-      },
-      {
-        incremental: [
-          {
-            data: { friends: [{ name: 'Han' }] },
-            path: ['hero'],
-          },
-        ],
-        hasNext: false,
-      },
-    ]);
+    expectJSON(result).toDeepEqual({
+      data: { hero: { friends: [{ name: 'Han' }] } },
+    });
   });
 
-  it('Handles empty async iterable list fields', async () => {
+  it('Deduplicates empty async iterable list fields', async () => {
     const document = parse(`
       query {
         hero {
@@ -2028,24 +2085,12 @@ describe('Execute: legacy defer directive format', () => {
         },
       },
     });
-    expectJSON(result).toDeepEqual([
-      {
-        data: { hero: { friends: [] } },
-        hasNext: true,
-      },
-      {
-        incremental: [
-          {
-            data: { friends: [] },
-            path: ['hero'],
-          },
-        ],
-        hasNext: false,
-      },
-    ]);
+    expectJSON(result).toDeepEqual({
+      data: { hero: { friends: [] } },
+    });
   });
 
-  it('Handles list fields with non-overlapping fields', async () => {
+  it('Does not deduplicate list fields with non-overlapping fields', async () => {
     const document = parse(`
       query {
         hero {
@@ -2068,23 +2113,34 @@ describe('Execute: legacy defer directive format', () => {
             friends: [{ name: 'Han' }, { name: 'Leia' }, { name: 'C-3PO' }],
           },
         },
+        pending: [{ id: '0', path: ['hero'] }],
         hasNext: true,
       },
       {
         incremental: [
           {
-            data: {
-              friends: [{ id: '2' }, { id: '3' }, { id: '4' }],
-            },
-            path: ['hero'],
+            data: { id: '2' },
+            id: '0',
+            subPath: ['friends', 0],
+          },
+          {
+            data: { id: '3' },
+            id: '0',
+            subPath: ['friends', 1],
+          },
+          {
+            data: { id: '4' },
+            id: '0',
+            subPath: ['friends', 2],
           },
         ],
+        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
   });
 
-  it('Handles list fields that return empty lists', async () => {
+  it('Deduplicates list fields that return empty lists', async () => {
     const document = parse(`
       query {
         hero {
@@ -2105,24 +2161,12 @@ describe('Execute: legacy defer directive format', () => {
         friends: () => [],
       },
     });
-    expectJSON(result).toDeepEqual([
-      {
-        data: { hero: { friends: [] } },
-        hasNext: true,
-      },
-      {
-        incremental: [
-          {
-            data: { friends: [] },
-            path: ['hero'],
-          },
-        ],
-        hasNext: false,
-      },
-    ]);
+    expectJSON(result).toDeepEqual({
+      data: { hero: { friends: [] } },
+    });
   });
 
-  it('Handles duplicated null object fields', async () => {
+  it('Deduplicates null object fields', async () => {
     const document = parse(`
       query {
         hero {
@@ -2143,24 +2187,12 @@ describe('Execute: legacy defer directive format', () => {
         nestedObject: () => null,
       },
     });
-    expectJSON(result).toDeepEqual([
-      {
-        data: { hero: { nestedObject: null } },
-        hasNext: true,
-      },
-      {
-        incremental: [
-          {
-            data: { nestedObject: null },
-            path: ['hero'],
-          },
-        ],
-        hasNext: false,
-      },
-    ]);
+    expectJSON(result).toDeepEqual({
+      data: { hero: { nestedObject: null } },
+    });
   });
 
-  it('Handles duplicated promise object fields', async () => {
+  it('Deduplicates promise object fields', async () => {
     const document = parse(`
       query {
         hero {
@@ -2180,21 +2212,9 @@ describe('Execute: legacy defer directive format', () => {
         nestedObject: () => Promise.resolve({ name: 'foo' }),
       },
     });
-    expectJSON(result).toDeepEqual([
-      {
-        data: { hero: { nestedObject: { name: 'foo' } } },
-        hasNext: true,
-      },
-      {
-        incremental: [
-          {
-            data: { nestedObject: { name: 'foo' } },
-            path: ['hero'],
-          },
-        ],
-        hasNext: false,
-      },
-    ]);
+    expectJSON(result).toDeepEqual({
+      data: { hero: { nestedObject: { name: 'foo' } } },
+    });
   });
 
   it('Handles errors thrown in deferred fragments', async () => {
@@ -2220,12 +2240,14 @@ describe('Execute: legacy defer directive format', () => {
     expectJSON(result).toDeepEqual([
       {
         data: { hero: { id: '1' } },
+        pending: [{ id: '0', path: ['hero'] }],
         hasNext: true,
       },
       {
         incremental: [
           {
             data: { name: null },
+            id: '0',
             errors: [
               {
                 message: 'bad',
@@ -2233,9 +2255,9 @@ describe('Execute: legacy defer directive format', () => {
                 path: ['hero', 'name'],
               },
             ],
-            path: ['hero'],
           },
         ],
+        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -2261,12 +2283,13 @@ describe('Execute: legacy defer directive format', () => {
     expectJSON(result).toDeepEqual([
       {
         data: { hero: { id: '1' } },
+        pending: [{ id: '0', path: ['hero'] }],
         hasNext: true,
       },
       {
-        incremental: [
+        completed: [
           {
-            data: null,
+            id: '0',
             errors: [
               {
                 message:
@@ -2275,7 +2298,6 @@ describe('Execute: legacy defer directive format', () => {
                 path: ['hero', 'nonNullName'],
               },
             ],
-            path: ['hero'],
           },
         ],
         hasNext: false,
@@ -2340,12 +2362,13 @@ describe('Execute: legacy defer directive format', () => {
     expectJSON(result).toDeepEqual([
       {
         data: { hero: { id: '1' } },
+        pending: [{ id: '0', path: ['hero'] }],
         hasNext: true,
       },
       {
-        incremental: [
+        completed: [
           {
-            data: null,
+            id: '0',
             errors: [
               {
                 message:
@@ -2354,7 +2377,6 @@ describe('Execute: legacy defer directive format', () => {
                 path: ['hero', 'nonNullName'],
               },
             ],
-            path: ['hero'],
           },
         ],
         hasNext: false,
@@ -2393,18 +2415,25 @@ describe('Execute: legacy defer directive format', () => {
         data: {
           hero: { id: '1' },
         },
+        pending: [{ id: '0', path: ['hero'] }],
         hasNext: true,
       },
       {
+        pending: [
+          { id: '1', path: ['hero', 'friends', 0] },
+          { id: '2', path: ['hero', 'friends', 1] },
+          { id: '3', path: ['hero', 'friends', 2] },
+        ],
         incremental: [
           {
             data: { name: 'slow', friends: [{}, {}, {}] },
-            path: ['hero'],
+            id: '0',
           },
-          { data: { name: 'Han' }, path: ['hero', 'friends', 0] },
-          { data: { name: 'Leia' }, path: ['hero', 'friends', 1] },
-          { data: { name: 'C-3PO' }, path: ['hero', 'friends', 2] },
+          { data: { name: 'Han' }, id: '1' },
+          { data: { name: 'Leia' }, id: '2' },
+          { data: { name: 'C-3PO' }, id: '3' },
         ],
+        completed: [{ id: '0' }, { id: '1' }, { id: '2' }, { id: '3' }],
         hasNext: false,
       },
     ]);
@@ -2433,21 +2462,28 @@ describe('Execute: legacy defer directive format', () => {
         data: {
           hero: { id: '1' },
         },
+        pending: [{ id: '0', path: ['hero'] }],
         hasNext: true,
       },
       {
+        pending: [
+          { id: '1', path: ['hero', 'friends', 0] },
+          { id: '2', path: ['hero', 'friends', 1] },
+          { id: '3', path: ['hero', 'friends', 2] },
+        ],
         incremental: [
           {
             data: {
               name: 'Luke',
               friends: [{}, {}, {}],
             },
-            path: ['hero'],
+            id: '0',
           },
-          { data: { name: 'Han' }, path: ['hero', 'friends', 0] },
-          { data: { name: 'Leia' }, path: ['hero', 'friends', 1] },
-          { data: { name: 'C-3PO' }, path: ['hero', 'friends', 2] },
+          { data: { name: 'Han' }, id: '1' },
+          { data: { name: 'Leia' }, id: '2' },
+          { data: { name: 'C-3PO' }, id: '3' },
         ],
+        completed: [{ id: '0' }, { id: '1' }, { id: '2' }, { id: '3' }],
         hasNext: false,
       },
     ]);
