+++ src/transform/__tests__/legacy-stream-test.ts
@@ -1,9 +1,5 @@
 import { assert, expect } from 'chai';
-import type {
-  DocumentNode,
-  InitialIncrementalExecutionResult,
-  SubsequentIncrementalExecutionResult,
-} from 'graphql';
+import type { DocumentNode } from 'graphql';
 import {
   GraphQLID,
   GraphQLList,
@@ -22,7 +18,11 @@ import { resolveOnNextTick } from '../../__testUtils__/resolveOnNextTick.js';
 import type { PromiseOrValue } from '../../jsutils/PromiseOrValue.js';
 import { promiseWithResolvers } from '../../jsutils/promiseWithResolvers.js';
 
-import { execute } from './execute.js';
+import type {
+  LegacyInitialIncrementalExecutionResult,
+  LegacySubsequentIncrementalExecutionResult,
+} from '../legacyExecuteIncrementally.js';
+import { legacyExecuteIncrementally as execute } from '../legacyExecuteIncrementally.js';
 
 const friendType = new GraphQLObjectType({
   fields: {
@@ -98,7 +98,8 @@ async function complete(
 
   if ('initialResult' in result) {
     const results: Array<
-      InitialIncrementalExecutionResult | SubsequentIncrementalExecutionResult
+      | LegacyInitialIncrementalExecutionResult
+      | LegacySubsequentIncrementalExecutionResult
     > = [result.initialResult];
     for await (const patch of result.subsequentResults) {
       results.push(patch);
@@ -126,7 +127,8 @@ async function completeAsync(
   const promises: Array<
     PromiseOrValue<
       IteratorResult<
-        InitialIncrementalExecutionResult | SubsequentIncrementalExecutionResult
+        | LegacyInitialIncrementalExecutionResult
+        | LegacySubsequentIncrementalExecutionResult
       >
     >
   > = [{ done: false, value: result.initialResult }];
@@ -136,7 +138,7 @@ async function completeAsync(
   return Promise.all(promises);
 }
 
-describe('Execute: stream directive', () => {
+describe('Execute: legacy stream directive', () => {
   it('Can stream a list field', async () => {
     const document = parse('{ scalarList @stream(initialCount: 1) }');
     const result = await complete(document, {
@@ -147,12 +149,12 @@ describe('Execute: stream directive', () => {
         data: {
           scalarList: ['apple'],
         },
-        pending: [{ id: '0', path: ['scalarList'] }],
         hasNext: true,
       },
       {
-        incremental: [{ items: ['banana', 'coconut'], id: '0' }],
-        completed: [{ id: '0' }],
+        incremental: [
+          { items: ['banana', 'coconut'], path: ['scalarList', 1] },
+        ],
         hasNext: false,
       },
     ]);
@@ -167,12 +169,12 @@ describe('Execute: stream directive', () => {
         data: {
           scalarList: [],
         },
-        pending: [{ id: '0', path: ['scalarList'] }],
         hasNext: true,
       },
       {
-        incremental: [{ items: ['apple', 'banana', 'coconut'], id: '0' }],
-        completed: [{ id: '0' }],
+        incremental: [
+          { items: ['apple', 'banana', 'coconut'], path: ['scalarList', 0] },
+        ],
         hasNext: false,
       },
     ]);
@@ -212,17 +214,16 @@ describe('Execute: stream directive', () => {
         data: {
           scalarList: ['apple'],
         },
-        pending: [{ id: '0', path: ['scalarList'], label: 'scalar-stream' }],
         hasNext: true,
       },
       {
         incremental: [
           {
             items: ['banana', 'coconut'],
-            id: '0',
+            path: ['scalarList', 1],
+            label: 'scalar-stream',
           },
         ],
-        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -248,12 +249,10 @@ describe('Execute: stream directive', () => {
     expectJSON(result).toDeepEqual([
       {
         data: { scalarList: ['apple', 'banana'] },
-        pending: [{ id: '0', path: ['scalarList'] }],
         hasNext: true,
       },
       {
-        incremental: [{ items: ['coconut'], id: '0' }],
-        completed: [{ id: '0' }],
+        incremental: [{ items: ['coconut'], path: ['scalarList', 2] }],
         hasNext: false,
       },
     ]);
@@ -272,7 +271,6 @@ describe('Execute: stream directive', () => {
         data: {
           scalarListList: [['apple', 'apple', 'apple']],
         },
-        pending: [{ id: '0', path: ['scalarListList'] }],
         hasNext: true,
       },
       {
@@ -282,10 +280,9 @@ describe('Execute: stream directive', () => {
               ['banana', 'banana', 'banana'],
               ['coconut', 'coconut', 'coconut'],
             ],
-            id: '0',
+            path: ['scalarListList', 1],
           },
         ],
-        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -316,7 +313,6 @@ describe('Execute: stream directive', () => {
             },
           ],
         },
-        pending: [{ id: '0', path: ['friendList'] }],
         hasNext: true,
       },
       {
@@ -328,10 +324,9 @@ describe('Execute: stream directive', () => {
                 id: '3',
               },
             ],
-            id: '0',
+            path: ['friendList', 2],
           },
         ],
-        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -353,14 +348,13 @@ describe('Execute: stream directive', () => {
         data: {
           friendList: [],
         },
-        pending: [{ id: '0', path: ['friendList'] }],
         hasNext: true,
       },
       {
         incremental: [
           {
             items: [{ name: 'Luke', id: '1' }],
-            id: '0',
+            path: ['friendList', 0],
           },
         ],
         hasNext: true,
@@ -369,7 +363,7 @@ describe('Execute: stream directive', () => {
         incremental: [
           {
             items: [{ name: 'Han', id: '2' }],
-            id: '0',
+            path: ['friendList', 1],
           },
         ],
         hasNext: true,
@@ -378,10 +372,9 @@ describe('Execute: stream directive', () => {
         incremental: [
           {
             items: [{ name: 'Leia', id: '3' }],
-            id: '0',
+            path: ['friendList', 2],
           },
         ],
-        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -414,14 +407,13 @@ describe('Execute: stream directive', () => {
         data: {
           friendList: [],
         },
-        pending: [{ id: '0', path: ['friendList'] }],
         hasNext: true,
       },
       {
         incremental: [
           {
             items: [{ id: '1' }],
-            id: '0',
+            path: ['friendList', 0],
           },
         ],
         hasNext: true,
@@ -430,7 +422,7 @@ describe('Execute: stream directive', () => {
         incremental: [
           {
             items: [{ id: '2' }],
-            id: '0',
+            path: ['friendList', 1],
           },
         ],
         hasNext: true,
@@ -439,10 +431,9 @@ describe('Execute: stream directive', () => {
         incremental: [
           {
             items: [{ id: '3' }],
-            id: '0',
+            path: ['friendList', 2],
           },
         ],
-        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -480,17 +471,15 @@ describe('Execute: stream directive', () => {
         data: {
           friendList: [],
         },
-        pending: [{ id: '0', path: ['friendList'] }],
         hasNext: true,
       },
       {
         incremental: [
           {
             items: [{ id: '1' }, { id: '2' }, { id: '3' }],
-            id: '0',
+            path: ['friendList', 0],
           },
         ],
-        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -526,7 +515,6 @@ describe('Execute: stream directive', () => {
             },
           ],
         },
-        pending: [{ id: '0', path: ['friendList'] }],
         hasNext: true,
       },
       {
@@ -538,10 +526,9 @@ describe('Execute: stream directive', () => {
                 id: '3',
               },
             ],
-            id: '0',
+            path: ['friendList', 2],
           },
         ],
-        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -576,17 +563,15 @@ describe('Execute: stream directive', () => {
         data: {
           friendList: [{ name: 'Luke', id: '1' }, null],
         },
-        pending: [{ id: '0', path: ['friendList'] }],
         hasNext: true,
       },
       {
         incremental: [
           {
             items: [{ name: 'Leia', id: '3' }],
-            id: '0',
+            path: ['friendList', 2],
           },
         ],
-        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -614,14 +599,12 @@ describe('Execute: stream directive', () => {
         data: {
           friendList: [{ name: 'Luke', id: '1' }],
         },
-        pending: [{ id: '0', path: ['friendList'] }],
         hasNext: true,
       },
       {
         incremental: [
           {
             items: [null],
-            id: '0',
             errors: [
               {
                 message: 'bad',
@@ -629,6 +612,7 @@ describe('Execute: stream directive', () => {
                 path: ['friendList', 1],
               },
             ],
+            path: ['friendList', 1],
           },
         ],
         hasNext: true,
@@ -637,10 +621,9 @@ describe('Execute: stream directive', () => {
         incremental: [
           {
             items: [{ name: 'Leia', id: '3' }],
-            id: '0',
+            path: ['friendList', 2],
           },
         ],
-        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -666,14 +649,13 @@ describe('Execute: stream directive', () => {
         data: {
           friendList: [],
         },
-        pending: [{ id: '0', path: ['friendList'] }],
         hasNext: true,
       },
       {
         incremental: [
           {
             items: [{ name: 'Luke', id: '1' }],
-            id: '0',
+            path: ['friendList', 0],
           },
         ],
         hasNext: true,
@@ -682,7 +664,7 @@ describe('Execute: stream directive', () => {
         incremental: [
           {
             items: [{ name: 'Han', id: '2' }],
-            id: '0',
+            path: ['friendList', 1],
           },
         ],
         hasNext: true,
@@ -691,13 +673,12 @@ describe('Execute: stream directive', () => {
         incremental: [
           {
             items: [{ name: 'Leia', id: '3' }],
-            id: '0',
+            path: ['friendList', 2],
           },
         ],
         hasNext: true,
       },
       {
-        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -726,20 +707,18 @@ describe('Execute: stream directive', () => {
             { name: 'Han', id: '2' },
           ],
         },
-        pending: [{ id: '0', path: ['friendList'] }],
         hasNext: true,
       },
       {
         incremental: [
           {
             items: [{ name: 'Leia', id: '3' }],
-            id: '0',
+            path: ['friendList', 2],
           },
         ],
         hasNext: true,
       },
       {
-        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -803,14 +782,13 @@ describe('Execute: stream directive', () => {
         data: {
           friendList: [],
         },
-        pending: [{ id: '0', path: ['friendList'] }],
         hasNext: true,
       },
       {
         incremental: [
           {
             items: [{ id: '1' }],
-            id: '0',
+            path: ['friendList', 0],
           },
         ],
         hasNext: true,
@@ -819,7 +797,7 @@ describe('Execute: stream directive', () => {
         incremental: [
           {
             items: [{ id: '2' }],
-            id: '0',
+            path: ['friendList', 1],
           },
         ],
         hasNext: true,
@@ -828,13 +806,12 @@ describe('Execute: stream directive', () => {
         incremental: [
           {
             items: [{ id: '3' }],
-            id: '0',
+            path: ['friendList', 2],
           },
         ],
         hasNext: true,
       },
       {
-        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -876,17 +853,15 @@ describe('Execute: stream directive', () => {
         data: {
           friendList: [],
         },
-        pending: [{ id: '0', path: ['friendList'] }],
         hasNext: true,
       },
       {
         incremental: [
           {
             items: [{ id: '1' }, { id: '2' }, { id: '3' }],
-            id: '0',
+            path: ['friendList', 0],
           },
         ],
-        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -918,7 +893,6 @@ describe('Execute: stream directive', () => {
               { name: 'Han', id: '2' },
             ],
           },
-          pending: [{ id: '0', path: ['friendList'] }],
           hasNext: true,
         },
       },
@@ -928,7 +902,7 @@ describe('Execute: stream directive', () => {
           incremental: [
             {
               items: [{ name: 'Leia', id: '3' }],
-              id: '0',
+              path: ['friendList', 2],
             },
           ],
           hasNext: true,
@@ -937,7 +911,6 @@ describe('Execute: stream directive', () => {
       {
         done: false,
         value: {
-          completed: [{ id: '0' }],
           hasNext: false,
         },
       },
@@ -992,13 +965,12 @@ describe('Execute: stream directive', () => {
         data: {
           friendList: [{ name: 'Luke', id: '1' }],
         },
-        pending: [{ id: '0', path: ['friendList'] }],
         hasNext: true,
       },
       {
-        completed: [
+        incremental: [
           {
-            id: '0',
+            items: null,
             errors: [
               {
                 message: 'bad',
@@ -1006,6 +978,7 @@ describe('Execute: stream directive', () => {
                 path: ['friendList'],
               },
             ],
+            path: ['friendList', 1],
           },
         ],
         hasNext: false,
@@ -1029,13 +1002,12 @@ describe('Execute: stream directive', () => {
         data: {
           nonNullFriendList: [{ name: 'Luke' }],
         },
-        pending: [{ id: '0', path: ['nonNullFriendList'] }],
         hasNext: true,
       },
       {
-        completed: [
+        incremental: [
           {
-            id: '0',
+            items: null,
             errors: [
               {
                 message:
@@ -1044,6 +1016,7 @@ describe('Execute: stream directive', () => {
                 path: ['nonNullFriendList', 1],
               },
             ],
+            path: ['nonNullFriendList', 1],
           },
         ],
         hasNext: false,
@@ -1076,13 +1049,12 @@ describe('Execute: stream directive', () => {
         data: {
           nonNullFriendList: [{ name: 'Luke' }],
         },
-        pending: [{ id: '0', path: ['nonNullFriendList'] }],
         hasNext: true,
       },
       {
-        completed: [
+        incremental: [
           {
-            id: '0',
+            items: null,
             errors: [
               {
                 message:
@@ -1091,6 +1063,7 @@ describe('Execute: stream directive', () => {
                 path: ['nonNullFriendList', 1],
               },
             ],
+            path: ['nonNullFriendList', 1],
           },
         ],
         hasNext: false,
@@ -1111,14 +1084,12 @@ describe('Execute: stream directive', () => {
         data: {
           scalarList: ['Luke'],
         },
-        pending: [{ id: '0', path: ['scalarList'] }],
         hasNext: true,
       },
       {
         incremental: [
           {
             items: [null],
-            id: '0',
             errors: [
               {
                 message: 'String cannot represent value: {}',
@@ -1126,9 +1097,9 @@ describe('Execute: stream directive', () => {
                 path: ['scalarList', 1],
               },
             ],
+            path: ['scalarList', 1],
           },
         ],
-        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -1155,14 +1126,12 @@ describe('Execute: stream directive', () => {
         data: {
           friendList: [{ nonNullName: 'Luke' }],
         },
-        pending: [{ id: '0', path: ['friendList'] }],
         hasNext: true,
       },
       {
         incremental: [
           {
             items: [null],
-            id: '0',
             errors: [
               {
                 message: 'Oops',
@@ -1170,6 +1139,7 @@ describe('Execute: stream directive', () => {
                 path: ['friendList', 1, 'nonNullName'],
               },
             ],
+            path: ['friendList', 1],
           },
         ],
         hasNext: true,
@@ -1178,10 +1148,9 @@ describe('Execute: stream directive', () => {
         incremental: [
           {
             items: [{ nonNullName: 'Han' }],
-            id: '0',
+            path: ['friendList', 2],
           },
         ],
-        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -1206,14 +1175,12 @@ describe('Execute: stream directive', () => {
         data: {
           friendList: [{ nonNullName: 'Luke' }],
         },
-        pending: [{ id: '0', path: ['friendList'] }],
         hasNext: true,
       },
       {
         incremental: [
           {
             items: [null],
-            id: '0',
             errors: [
               {
                 message: 'Oops',
@@ -1221,6 +1188,7 @@ describe('Execute: stream directive', () => {
                 path: ['friendList', 1, 'nonNullName'],
               },
             ],
+            path: ['friendList', 1],
           },
         ],
         hasNext: true,
@@ -1229,10 +1197,9 @@ describe('Execute: stream directive', () => {
         incremental: [
           {
             items: [{ nonNullName: 'Han' }],
-            id: '0',
+            path: ['friendList', 2],
           },
         ],
-        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -1259,13 +1226,12 @@ describe('Execute: stream directive', () => {
         data: {
           nonNullFriendList: [{ nonNullName: 'Luke' }],
         },
-        pending: [{ id: '0', path: ['nonNullFriendList'] }],
         hasNext: true,
       },
       {
-        completed: [
+        incremental: [
           {
-            id: '0',
+            items: null,
             errors: [
               {
                 message: 'Oops',
@@ -1273,6 +1239,7 @@ describe('Execute: stream directive', () => {
                 path: ['nonNullFriendList', 1, 'nonNullName'],
               },
             ],
+            path: ['nonNullFriendList', 1],
           },
         ],
         hasNext: false,
@@ -1299,13 +1266,12 @@ describe('Execute: stream directive', () => {
         data: {
           nonNullFriendList: [{ nonNullName: 'Luke' }],
         },
-        pending: [{ id: '0', path: ['nonNullFriendList'] }],
         hasNext: true,
       },
       {
-        completed: [
+        incremental: [
           {
-            id: '0',
+            items: null,
             errors: [
               {
                 message: 'Oops',
@@ -1313,6 +1279,7 @@ describe('Execute: stream directive', () => {
                 path: ['nonNullFriendList', 1, 'nonNullName'],
               },
             ],
+            path: ['nonNullFriendList', 1],
           },
         ],
         hasNext: false,
@@ -1341,14 +1308,12 @@ describe('Execute: stream directive', () => {
         data: {
           friendList: [{ nonNullName: 'Luke' }],
         },
-        pending: [{ id: '0', path: ['friendList'] }],
         hasNext: true,
       },
       {
         incremental: [
           {
             items: [null],
-            id: '0',
             errors: [
               {
                 message: 'Oops',
@@ -1356,6 +1321,7 @@ describe('Execute: stream directive', () => {
                 path: ['friendList', 1, 'nonNullName'],
               },
             ],
+            path: ['friendList', 1],
           },
         ],
         hasNext: true,
@@ -1364,13 +1330,12 @@ describe('Execute: stream directive', () => {
         incremental: [
           {
             items: [{ nonNullName: 'Han' }],
-            id: '0',
+            path: ['friendList', 2],
           },
         ],
         hasNext: true,
       },
       {
-        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -1396,13 +1361,12 @@ describe('Execute: stream directive', () => {
         data: {
           nonNullFriendList: [{ nonNullName: 'Luke' }],
         },
-        pending: [{ id: '0', path: ['nonNullFriendList'] }],
         hasNext: true,
       },
       {
-        completed: [
+        incremental: [
           {
-            id: '0',
+            items: null,
             errors: [
               {
                 message: 'Oops',
@@ -1410,6 +1374,7 @@ describe('Execute: stream directive', () => {
                 path: ['nonNullFriendList', 1, 'nonNullName'],
               },
             ],
+            path: ['nonNullFriendList', 1],
           },
         ],
         hasNext: false,
@@ -1459,13 +1424,12 @@ describe('Execute: stream directive', () => {
         data: {
           nonNullFriendList: [{ nonNullName: 'Luke' }],
         },
-        pending: [{ id: '0', path: ['nonNullFriendList'] }],
         hasNext: true,
       },
       {
-        completed: [
+        incremental: [
           {
-            id: '0',
+            items: null,
             errors: [
               {
                 message: 'Oops',
@@ -1473,6 +1437,7 @@ describe('Execute: stream directive', () => {
                 path: ['nonNullFriendList', 1, 'nonNullName'],
               },
             ],
+            path: ['nonNullFriendList', 1],
           },
         ],
         hasNext: false,
@@ -1532,13 +1497,12 @@ describe('Execute: stream directive', () => {
         data: {
           nonNullFriendList: [{ nonNullName: 'Luke' }],
         },
-        pending: [{ id: '0', path: ['nonNullFriendList'] }],
         hasNext: true,
       },
       {
-        completed: [
+        incremental: [
           {
-            id: '0',
+            items: null,
             errors: [
               {
                 message: 'Oops',
@@ -1546,6 +1510,7 @@ describe('Execute: stream directive', () => {
                 path: ['nonNullFriendList', 1, 'nonNullName'],
               },
             ],
+            path: ['nonNullFriendList', 1],
           },
         ],
         hasNext: false,
@@ -1648,21 +1613,16 @@ describe('Execute: stream directive', () => {
           otherNestedObject: {},
           nestedObject: { nestedFriendList: [] },
         },
-        pending: [
-          { id: '0', path: ['otherNestedObject'] },
-          { id: '1', path: ['nestedObject', 'nestedFriendList'] },
-        ],
         hasNext: true,
       },
       {
         incremental: [
           {
             items: [{ name: 'Luke' }],
-            id: '1',
+            path: ['nestedObject', 'nestedFriendList', 0],
           },
           {
             data: { scalarField: null },
-            id: '0',
             errors: [
               {
                 message: 'Oops',
@@ -1670,13 +1630,12 @@ describe('Execute: stream directive', () => {
                 path: ['otherNestedObject', 'scalarField'],
               },
             ],
+            path: ['otherNestedObject'],
           },
         ],
-        completed: [{ id: '0' }],
         hasNext: true,
       },
       {
-        completed: [{ id: '1' }],
         hasNext: false,
       },
     ]);
@@ -1711,7 +1670,6 @@ describe('Execute: stream directive', () => {
         data: {
           nestedObject: {},
         },
-        pending: [{ id: '0', path: ['nestedObject'] }],
         hasNext: true,
       },
       {
@@ -1720,7 +1678,6 @@ describe('Execute: stream directive', () => {
             data: {
               deeperNestedObject: null,
             },
-            id: '0',
             errors: [
               {
                 message:
@@ -1733,9 +1690,9 @@ describe('Execute: stream directive', () => {
                 ],
               },
             ],
+            path: ['nestedObject'],
           },
         ],
-        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -1764,14 +1721,12 @@ describe('Execute: stream directive', () => {
         data: {
           friendList: [],
         },
-        pending: [{ id: '0', path: ['friendList'] }],
         hasNext: true,
       },
       {
         incremental: [
           {
             items: [null],
-            id: '0',
             errors: [
               {
                 message:
@@ -1780,12 +1735,12 @@ describe('Execute: stream directive', () => {
                 path: ['friendList', 0, 'nonNullName'],
               },
             ],
+            path: ['friendList', 0],
           },
         ],
         hasNext: true,
       },
       {
-        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -1856,7 +1811,6 @@ describe('Execute: stream directive', () => {
       data: {
         nestedObject: {},
       },
-      pending: [{ id: '0', path: ['nestedObject'] }],
       hasNext: true,
     });
 
@@ -1869,7 +1823,6 @@ describe('Execute: stream directive', () => {
             data: {
               deeperNestedObject: null,
             },
-            id: '0',
             errors: [
               {
                 message:
@@ -1882,9 +1835,9 @@ describe('Execute: stream directive', () => {
                 ],
               },
             ],
+            path: ['nestedObject'],
           },
         ],
-        completed: [{ id: '0' }],
         hasNext: false,
       },
     });
@@ -1918,14 +1871,13 @@ describe('Execute: stream directive', () => {
         data: {
           friendList: [{ id: '1', name: 'Luke' }],
         },
-        pending: [{ id: '0', path: ['friendList'] }],
         hasNext: true,
       },
       {
         incremental: [
           {
             items: [{ id: '2', name: 'Han' }],
-            id: '0',
+            path: ['friendList', 1],
           },
         ],
         hasNext: true,
@@ -1934,13 +1886,12 @@ describe('Execute: stream directive', () => {
         incremental: [
           {
             items: [{ id: '3', name: 'Leia' }],
-            id: '0',
+            path: ['friendList', 2],
           },
         ],
         hasNext: true,
       },
       {
-        completed: [{ id: '0' }],
         hasNext: false,
       },
     ]);
@@ -1978,14 +1929,26 @@ describe('Execute: stream directive', () => {
             nestedFriendList: [],
           },
         },
-        pending: [{ id: '0', path: ['nestedObject', 'nestedFriendList'] }],
         hasNext: true,
       },
       {
         incremental: [
+          {
+            data: { nestedFriendList: [] },
+            path: ['nestedObject'],
+          },
+        ],
+        hasNext: true,
+      },
+      {
+        incremental: [
+          {
+            items: [{ id: '1' }],
+            path: ['nestedObject', 'nestedFriendList', 0],
+          },
           {
             items: [{ id: '1', name: 'Luke' }],
-            id: '0',
+            path: ['nestedObject', 'nestedFriendList', 0],
           },
         ],
         hasNext: true,
@@ -1993,14 +1956,155 @@ describe('Execute: stream directive', () => {
       {
         incremental: [
           {
+            items: [{ id: '2' }],
+            path: ['nestedObject', 'nestedFriendList', 1],
+          },
+          {
             items: [{ id: '2', name: 'Han' }],
-            id: '0',
+            path: ['nestedObject', 'nestedFriendList', 1],
           },
         ],
         hasNext: true,
       },
       {
-        completed: [{ id: '0' }],
+        hasNext: false,
+      },
+    ]);
+  });
+  it('Handles overlapping deferred and non-deferred streams with a non-zero initial count', async () => {
+    const document = parse(`
+      query {
+        nestedObject {
+          nestedFriendList @stream(initialCount: 1) {
+            id
+          }
+        }
+        nestedObject {
+          ... @defer {
+            nestedFriendList @stream(initialCount: 1) {
+              id
+              name
+            }
+          }
+        }
+      }
+    `);
+    const result = await complete(document, {
+      nestedObject: {
+        async *nestedFriendList() {
+          yield await Promise.resolve(friends[0]);
+          yield await Promise.resolve(friends[1]);
+        },
+      },
+    });
+    expectJSON(result).toDeepEqual([
+      {
+        data: {
+          nestedObject: {
+            nestedFriendList: [{ id: '1' }],
+          },
+        },
+        hasNext: true,
+      },
+      {
+        incremental: [
+          {
+            data: { nestedFriendList: [{ id: '1', name: 'Luke' }] },
+            path: ['nestedObject'],
+          },
+        ],
+        hasNext: true,
+      },
+      {
+        incremental: [
+          {
+            items: [{ id: '2' }],
+            path: ['nestedObject', 'nestedFriendList', 1],
+          },
+          {
+            items: [{ id: '2', name: 'Han' }],
+            path: ['nestedObject', 'nestedFriendList', 1],
+          },
+        ],
+        hasNext: true,
+      },
+      {
+        hasNext: false,
+      },
+    ]);
+  });
+  it('Handles overlapping deferred and non-deferred streams with a non-zero initial count with a slow defer', async () => {
+    const document = parse(`
+      query {
+        nestedObject {
+          nestedFriendList @stream(initialCount: 1) {
+            id
+          }
+        }
+        nestedObject {
+          ... @defer {
+            nestedFriendList @stream(initialCount: 1) {
+              id
+              name
+            }
+          }
+        }
+      }
+    `);
+    const result = await complete(document, {
+      nestedObject: {
+        async *nestedFriendList() {
+          yield await Promise.resolve({
+            ...friends[0],
+            name: Promise.resolve(friends[0].name),
+          });
+          yield await Promise.resolve(friends[1]);
+          yield await Promise.resolve(friends[2]);
+        },
+      },
+    });
+    expectJSON(result).toDeepEqual([
+      {
+        data: {
+          nestedObject: {
+            nestedFriendList: [{ id: '1' }],
+          },
+        },
+        hasNext: true,
+      },
+      {
+        incremental: [
+          {
+            items: [{ id: '2' }],
+            path: ['nestedObject', 'nestedFriendList', 1],
+          },
+          {
+            data: {
+              nestedFriendList: [{ id: '1', name: 'Luke' }],
+            },
+            path: ['nestedObject'],
+          },
+          {
+            items: [{ id: '2', name: 'Han' }],
+            path: ['nestedObject', 'nestedFriendList', 1],
+          },
+        ],
+        hasNext: true,
+      },
+      {
+        incremental: [
+          {
+            items: [{ id: '3' }],
+            path: ['nestedObject', 'nestedFriendList', 2],
+          },
+          {
+            items: [{ id: '3', name: 'Leia' }],
+            path: ['nestedObject', 'nestedFriendList', 2],
+          },
+        ],
+        hasNext: true,
+      },
+      {
         hasNext: false,
       },
     ]);
@@ -2042,7 +2146,6 @@ describe('Execute: stream directive', () => {
       data: {
         nestedObject: {},
       },
-      pending: [{ id: '0', path: ['nestedObject'] }],
       hasNext: true,
     });
 
@@ -2051,14 +2154,12 @@ describe('Execute: stream directive', () => {
     const result2 = await result2Promise;
     expectJSON(result2).toDeepEqual({
       value: {
-        pending: [{ id: '1', path: ['nestedObject', 'nestedFriendList'] }],
         incremental: [
           {
             data: { scalarField: 'slow', nestedFriendList: [] },
-            id: '0',
+            path: ['nestedObject'],
           },
         ],
-        completed: [{ id: '0' }],
         hasNext: true,
       },
       done: false,
@@ -2070,7 +2171,7 @@ describe('Execute: stream directive', () => {
         incremental: [
           {
             items: [{ name: 'Luke' }],
-            id: '1',
+            path: ['nestedObject', 'nestedFriendList', 0],
           },
         ],
         hasNext: true,
@@ -2084,7 +2185,7 @@ describe('Execute: stream directive', () => {
         incremental: [
           {
             items: [{ name: 'Han' }],
-            id: '1',
+            path: ['nestedObject', 'nestedFriendList', 1],
           },
         ],
         hasNext: true,
@@ -2095,7 +2196,6 @@ describe('Execute: stream directive', () => {
     const result5 = await iterator.next();
     expectJSON(result5).toDeepEqual({
       value: {
-        completed: [{ id: '1' }],
         hasNext: false,
       },
       done: false,
@@ -2148,7 +2248,6 @@ describe('Execute: stream directive', () => {
       data: {
         friendList: [],
       },
-      pending: [{ id: '0', path: ['friendList'], label: 'stream-label' }],
       hasNext: true,
     });
 
@@ -2157,18 +2256,18 @@ describe('Execute: stream directive', () => {
     const result2 = await result2Promise;
     expectJSON(result2).toDeepEqual({
       value: {
-        pending: [{ id: '1', path: ['friendList', 0], label: 'DeferName' }],
         incremental: [
           {
             items: [{ id: '1' }],
-            id: '0',
+            path: ['friendList', 0],
+            label: 'stream-label',
           },
           {
             data: { name: 'Luke' },
-            id: '1',
+            path: ['friendList', 0],
+            label: 'DeferName',
           },
         ],
-        completed: [{ id: '1' }],
         hasNext: true,
       },
       done: false,
@@ -2179,11 +2278,11 @@ describe('Execute: stream directive', () => {
     const result3 = await result3Promise;
     expectJSON(result3).toDeepEqual({
       value: {
-        pending: [{ id: '2', path: ['friendList', 1], label: 'DeferName' }],
         incremental: [
           {
             items: [{ id: '2' }],
-            id: '0',
+            path: ['friendList', 1],
+            label: 'stream-label',
           },
         ],
         hasNext: true,
@@ -2193,27 +2292,19 @@ describe('Execute: stream directive', () => {
     const result4 = await iterator.next();
     expectJSON(result4).toDeepEqual({
       value: {
-        completed: [{ id: '0' }],
-        hasNext: true,
+        incremental: [
+          {
+            data: { name: 'Han' },
+            path: ['friendList', 1],
+            label: 'DeferName',
+          },
+        ],
+        hasNext: false,
       },
       done: false,
     });
     const result5 = await iterator.next();
     expectJSON(result5).toDeepEqual({
-      value: {
-        incremental: [
-          {
-            data: { name: 'Han' },
-            id: '2',
-          },
-        ],
-        completed: [{ id: '2' }],
-        hasNext: false,
-      },
-      done: false,
-    });
-    const result6 = await iterator.next();
-    expectJSON(result6).toDeepEqual({
       value: undefined,
       done: true,
     });
@@ -2260,10 +2351,6 @@ describe('Execute: stream directive', () => {
       data: {
         friendList: [{ id: '1' }],
       },
-      pending: [
-        { id: '0', path: ['friendList', 0], label: 'DeferName' },
-        { id: '1', path: ['friendList'], label: 'stream-label' },
-      ],
       hasNext: true,
     });
 
@@ -2275,10 +2362,10 @@ describe('Execute: stream directive', () => {
         incremental: [
           {
             data: { name: 'Luke' },
-            id: '0',
+            path: ['friendList', 0],
+            label: 'DeferName',
           },
         ],
-        completed: [{ id: '0' }],
         hasNext: true,
       },
       done: false,
@@ -2287,11 +2374,11 @@ describe('Execute: stream directive', () => {
     const result3 = await iterator.next();
     expectJSON(result3).toDeepEqual({
       value: {
-        pending: [{ id: '2', path: ['friendList', 1], label: 'DeferName' }],
         incremental: [
           {
             items: [{ id: '2' }],
-            id: '1',
+            path: ['friendList', 1],
+            label: 'stream-label',
           },
         ],
         hasNext: true,
@@ -2305,10 +2392,10 @@ describe('Execute: stream directive', () => {
         incremental: [
           {
             data: { name: 'Han' },
-            id: '2',
+            path: ['friendList', 1],
+            label: 'DeferName',
           },
         ],
-        completed: [{ id: '2' }],
         hasNext: true,
       },
       done: false,
@@ -2319,7 +2406,6 @@ describe('Execute: stream directive', () => {
     const result5 = await result5Promise;
     expectJSON(result5).toDeepEqual({
       value: {
-        completed: [{ id: '1' }],
         hasNext: false,
       },
       done: false,
@@ -2368,7 +2454,6 @@ describe('Execute: stream directive', () => {
       data: {
         friendList: [],
       },
-      pending: [{ id: '0', path: ['friendList'] }],
       hasNext: true,
     });
 
@@ -2426,7 +2511,6 @@ describe('Execute: stream directive', () => {
           },
         ],
       },
-      pending: [{ id: '0', path: ['friendList'] }],
       hasNext: true,
     });
 
@@ -2486,10 +2570,6 @@ describe('Execute: stream directive', () => {
           },
         ],
       },
-      pending: [
-        { id: '0', path: ['friendList', 0] },
-        { id: '1', path: ['friendList'] },
-      ],
       hasNext: true,
     });
